Update Main.py
i bound the enter button to the game.
problem with recalling the hangman game: i think the issue is with calling the category part when putting hangman(category, difficulty). I created a list called "pressed_buttons" to track where the user click and tried to input the values into the hangman function but when I click next button, it just shows an empty screen.

when you press enter after finishing the hangman, itll bring you to the next page